requirements.txt
runpod>=1.6.0
marker-pdf>=1.10.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
